Fix relative paths not working
When UHLSLMaterialFunctionLibrary::OnDirectoryChanged is called array
of FFileChangeData has paths that are always full, not relative.

Filepath in HLSLMaterialFunctionLibrary is stored as relative filepath
when possible. Which is good as this allowes for having shared hlsl
directory in the project for all project users.

But when OnDirectoryChanged is called with those full paths it fails to
find matching files.

With this change GetFilePath always returns full, not relative path