Update function on included file changes
Added watchers for files used in IncludeFilePaths.
I tried to keep the Bind/Unbind watcher logic as unchanged as possible but to
allow for reuse in the includes code.

Needed to add hash of their contents to hash so we know that contents of the
function really did change even if it's not visible in the body.

Needed to add hash of the function into the custom node code otherwise unreal
was being smart about it and wasn't recompiling the node.

Moved function hashing to earlier step and saved results as we need to use it
multiple times.