Resolve local includes
When using IDEs like Rider for developing shaders they often have
problems with finding includes when included as virtual paths.

See for example https://youtrack.jetbrains.com/issue/RIDER-69615

This can be often worked around when working with global shaders
by using relative paths to virtual shader location.

This cannot be done same way for code generated with HLSLMaterial.

This change adds fallback mechanism when we find that include path
isn't real virtual path.
We try to check if this path is relative path to our hlsl file and
if it is we try to find it in virtual mappings.