Correction - previous commit was not whitespace - notes below:
The point of that previous checkin was to add an include search
path for the shader compiler.  This is needed for the HLSL in the
custom material if you want to just include the file and call the
functions directly.  In the custom shader itself, this path is seen
as "/Project".  On disk, within the project, it is seen as
"/Shaders".